data_dashboard/backend/requirements.txt
# Backend dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.0
python-multipart==0.0.12
websockets==13.1
aiofiles==24.1.0

# Data processing
pandas==2.2.3
numpy==2.1.3

# Monitoring and logging
loguru==0.7.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2